backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
azure-storage-blob==12.23.0
azure-identity==1.17.1
azure-search-documents==11.6.0b5
langchain==0.2.11
langchain-openai==0.1.22
langchain-community==0.2.10
langchain-text-splitters==0.2.2
pypdf==4.3.1
pydantic==2.9.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
